requirements.txt
#_requirements.txt
discord.py
flask
gunicorn
supabase
PyJWT
requests>=2.28.0
